added anotehr test to common deeplearning tests
